feat(rediscovery): add updateExistingDevice function
